feat(notification): sync with launcher
