+++ shell/crates/launcher/Cargo.toml
@@ -19,6 +19,7 @@ status_bar = { path = "../status_bar" }
 settings_drawer = { path = "../settings_drawer" }
 universal_search = { path = "../universal_search" }
 running_apps = { path = "../running_apps" }
+notifications = { path = "../notifications" }
 homescreen = { path = "../homescreen" }
 dispatcher = { path = "../dispatcher" }
 shell_state = { path = "../shell_state" }
