+++ shell/crates/launcher/src/lib.rs
@@ -21,9 +21,11 @@ pub fn run() {
 
         running_apps::run_app(cx);
 
-        universal_search::run_app(cx);
+        // universal_search::run_app(cx);
 
         settings_drawer::run_app(cx);
+
+        notifications::run_app(cx);
 
         cx.activate(true);
         cx.refresh_windows();
