+++ shell/crates/notifications/Cargo.toml
@@ -14,6 +14,7 @@ gpui = { workspace = true }
 futures = { workspace = true }
 desktop_dbus = { path = "../../../dbus/mechanix/desktop"}
 commons = { path = "../commons"}
+settings = { path = "../settings"}
 smol = "2.0.2"
 uuid = "1.18.1"
 tracing = "0.1.41"
