+++ shell/crates/notifications/src/widgets/mod.rs
@@ -1,0 +1,18 @@
+mod center;
+mod notification;
+
+pub use center::{
+    NotificationList,
+    NotificationCenter,
+    NotificationWidget,
+};
+
+pub use notification::{
+    DbNotification,
+    UserDismissedEvent,
+    NotificationUi,
+};
+
+pub mod prelude {
+    pub use super::*;
+}
