+++ shell/crates/settings/src/settings.rs
@@ -17,6 +17,8 @@ pub struct Settings {
     #[serde(default)]
     pub universal_search: UniversalSearchSettings,
     #[serde(default)]
+    pub notifications: NotificationSettings,
+    #[serde(default)]
     pub app_drawer: AppDrawerSettings,
     #[serde(default)]
     pub homescreen: HomescreenSettings,
@@ -133,6 +135,30 @@ impl Default for UniversalSearchSettings {
                 layer: Layer::Top,
                 anchor: Anchor::BOTTOM,
                 namespace: "mechanix.universal.search".into(),
+                exclusive_zone: px(-1.0),
+                size: Size::new(px(540.0), px(620.0)),
+            },
+            navbar_size: Size::new(px(180.0), px(29.0)),
+        }
+    }
+}
+
+/// Notification settings
+#[derive(Debug, Clone, Deserialize, PartialEq)]
+pub struct NotificationSettings {
+    #[serde(default)]
+    pub layer_shell: LayerShellSettings,
+    #[serde(default)]
+    pub navbar_size: Size<Pixels>,
+}
+
+impl Default for NotificationSettings {
+    fn default() -> Self {
+        Self {
+            layer_shell: LayerShellSettings {
+                layer: Layer::Top,
+                anchor: Anchor::BOTTOM,
+                namespace: "mechanix.notifications".into(),
                 exclusive_zone: px(-1.0),
                 size: Size::new(px(540.0), px(620.0)),
             },
